src/requirements/base.txt
amqp==5.1.1
anyio==4.0.0
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
boto3==1.28.69
botocore==1.31.69
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
Django==4.2.6
django-cors-headers==4.3.0
django-extensions==3.2.3
djangorestframework==3.14.0
dnspython==2.4.2
drf-spectacular==0.26.5
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.3.2
packaging==23.2
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymongo==4.5.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
resend==0.6.0
rpds-py==0.10.6
s3transfer==0.7.0
sentry-sdk==1.32.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
zope.interface==6.1
